backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-multipart==0.0.16
pillow==11.0.0
numpy==2.2.4
opencv-python==4.10.0.84
